rename pages to portal/application
