+++ src/main/www/src/components/Application/CompanyInformation/CompanyInformation.js
