+++ src/main/www/src/components/Application/CompanyInformation/CompanyInformationVAT.js
