+++ src/main/www/src/components/Application/WorkingGroups/WorkingGroup.js
