+++ src/main/www/src/components/UIComponents/Templates/AppFooter.js
