+++ src/main/www/src/components/UIComponents/Templates/AppTemplate.js
